moved getting order notices to dashboard component
