+++ src/components/dashboardComponent.jsx
@@ -1,10 +1,34 @@
 import React from 'react';
 import {Grid, Row} from 'react-bootstrap';
 import OrderNotice from './orderNoticeComponent';
+import {connect} from 'react-redux';
+import {hydrateOrderNotices} from '../store/orderNotices';
 
 const Dashboard = React.createClass({
     propTypes: {
+        dispatch: React.PropTypes.func.isRequired,
         orderNotices: React.PropTypes.array.isRequired
+    },
+
+    componentWillMount () {
+        const mockedOrderNotices = [
+            {
+                id: 1,
+                restaurant: 'hello',
+                deliveryTime: '15:00',
+                orderer: 'Andrzej',
+                hungryGuysCount: 4
+            },
+            {
+                id: 2,
+                restaurant: 'world',
+                deliveryTime: '12:00',
+                orderer: 'Elo',
+                hungryGuysCount: 2
+            }
+        ];
+
+        this.props.dispatch(hydrateOrderNotices(mockedOrderNotices));
     },
 
     _renderOrderNotices () {
@@ -24,4 +48,8 @@ const Dashboard = React.createClass({
     }
 });
 
-export default Dashboard;
+const ConnectedDashboard = connect(
+    (state) => ({orderNotices: state.orderNotices})
+)(Dashboard);
+
+export default ConnectedDashboard;
