+++ src/components/index.jsx
@@ -1,37 +1,9 @@
 import React from 'react';
-import {connect} from 'react-redux';
-import {hydrateOrderNotices} from '../store/orderNotices';
 import {Grid, Row, Col, PageHeader} from 'react-bootstrap';
 
 import Dashboard from './dashboardComponent';
 
 const Index = React.createClass({
-    propTypes: {
-        dispatch: React.PropTypes.func.isRequired,
-        orderNotices: React.PropTypes.array.isRequired
-    },
-
-    componentWillMount () {
-        const mockedOrderNotices = [
-            {
-                id: 1,
-                restaurant: 'hello',
-                deliveryTime: '15:00',
-                orderer: 'Andrzej',
-                hungryGuysCount: 4
-            },
-            {
-                id: 2,
-                restaurant: 'world',
-                deliveryTime: '12:00',
-                orderer: 'Elo',
-                hungryGuysCount: 2
-            }
-        ];
-
-        this.props.dispatch(hydrateOrderNotices(mockedOrderNotices));
-    },
-
     render () {
         return (
             <Grid>
@@ -42,7 +14,7 @@ const Index = React.createClass({
                 </Row>
                 <Row>
                     <Col xs={12}>
-                        <Dashboard orderNotices={this.props.orderNotices} />
+                        <Dashboard />
                     </Col>
                 </Row>
             </Grid>
@@ -50,8 +22,4 @@ const Index = React.createClass({
     }
 });
 
-const ConnectedIndex = connect(
-    (state) => ({orderNotices: state.orderNotices})
-)(Index);
-
-export default ConnectedIndex;
+export default Index;
